+++ packages/LibTests/src/tests/compositionPipelineTests.ts
@@ -599,7 +599,7 @@ category('CompositionPipeline composition config', async () => {
       ...sconfig2,
       nestedPipelinesConfig: {
         'testPipeline1': {
-	  insertBeforeStep: 'step2',
+          insertBeforeStep: 'step2',
         },
       },
     };